moved interfaces to implementation namespaces
