+++ src/ExtractorInterface.php
@@ -1,17 +1,16 @@
 <?php
 namespace Thunder\Shortcode;
 
+use Thunder\Shortcode\Extractor\ExtractorInterface as ExtractorInterfaceBase;
+
 /**
- * @author Tomasz Kowalczyk <tomasz@kowalczyk.cc>
+ * This implementation is left only to not break IDE autocompletion, this class
+ * is deprecated, it was moved to the new location as specified in docblock.
+ * This file will be removed in version 1.0!
+ *
+ * @deprecated use Thunder\Shortcode\Extractor\RegexExtractor
+ * @codeCoverageIgnore
  */
-interface ExtractorInterface
+interface ExtractorInterface extends ExtractorInterfaceBase
     {
-    /**
-     * Extract shortcode string matches with their offsets for further analysis
-     *
-     * @param string $text Text to extract from
-     *
-     * @return Match[]
-     */
-    public function extract($text);
     }
