+++ src/Parser/ParserInterface.php
@@ -1,0 +1,19 @@
+<?php
+namespace Thunder\Shortcode\Parser;
+
+use Thunder\Shortcode\Shortcode;
+
+/**
+ * @author Tomasz Kowalczyk <tomasz@kowalczyk.cc>
+ */
+interface ParserInterface
+    {
+    /**
+     * Parse single shortcode match into object
+     *
+     * @param string $text
+     *
+     * @return Shortcode
+     */
+    public function parse($text);
+    }
