+++ src/ParserInterface.php
@@ -1,17 +1,16 @@
 <?php
 namespace Thunder\Shortcode;
 
+use Thunder\Shortcode\Parser\ParserInterface as ParserInterfaceBase;
+
 /**
- * @author Tomasz Kowalczyk <tomasz@kowalczyk.cc>
+ * This implementation is left only to not break IDE autocompletion, this class
+ * is deprecated, it was moved to the new location as specified in docblock.
+ * This file will be removed in version 1.0!
+ *
+ * @deprecated use Thunder\Shortcode\Extractor\RegexExtractor
+ * @codeCoverageIgnore
  */
-interface ParserInterface
+interface ParserInterface extends ParserInterfaceBase
     {
-    /**
-     * Parse single shortcode match into object
-     *
-     * @param string $text
-     *
-     * @return Shortcode
-     */
-    public function parse($text);
     }
