+++ src/Shortcode.php
@@ -1,6 +1,8 @@
 <?php
 namespace Thunder\Shortcode;
 
+use Thunder\Shortcode\Shortcode\ShortcodeInterface;
+
 /**
  * @author Tomasz Kowalczyk <tomasz@kowalczyk.cc>
  */
